Moved renderer to own file. Added colors to styles.css
