+++ src/processors/chat.ts
@@ -1,4 +1,5 @@
-import {MarkdownPostProcessorContext, MarkdownRenderChild, MarkdownRenderer, Plugin} from "obsidian";
+import {MarkdownPostProcessorContext, Plugin} from "obsidian";
+import {renderChat} from "../renderer";
 import {ChatComment, ChatDelimeter, ChatMessage, ChatNode} from "../types";
 
 abstract class Patterns {
@@ -27,35 +28,7 @@ export async function chatHandler(plugin: Plugin, src: string, el: HTMLElement, 
       nodes.push(<ChatComment>{kind: "comment", body: group.text ?? ""});
     }
   }
-  await renderChat(plugin, nodes, el, ctx);
-}
-
-async function renderChat(plugin: Plugin, nodes: ChatNode[], el: HTMLElement, ctx: MarkdownPostProcessorContext) {
-  const chatViewDiv = el.createDiv({cls: ["chat-view"]});
-  for (const node of nodes) {
-    if (node.kind === "message") {
-      const messageDiv = chatViewDiv.createDiv({cls: ["chat-view-message", `chat-view-message-${node.side}`]});
-      if (node.header.length > 0) {
-        messageDiv.createDiv({cls: ["chat-view-header"], text: node.header});
-      }
-      if (node.body.length > 0) {
-        const bodyDiv = messageDiv.createDiv({cls: ["chat-view-body"]});
-        const child = new MarkdownRenderChild(bodyDiv);
-        ctx.addChild(child);
-        await MarkdownRenderer.render(plugin.app, node.body, bodyDiv, ctx.sourcePath, child);
-      }
-      if (node.subtext.length > 0) {
-        messageDiv.createDiv({cls: ["chat-view-subtext"], text: node.subtext});
-      }
-    } else if (node.kind === "comment") {
-      chatViewDiv.createDiv({cls: ["chat-view-comment"], text: node.body});
-    } else if (node.kind === "delimeter") {
-      const delimeter = chatViewDiv.createDiv({cls: ["chat-view-delimeter"]});
-      for (let i = 0; i < 3; i++) {
-        delimeter.createDiv({cls: ["chat-view-delimeter-dot"]});
-      }
-    }
-  }
+  await renderChat(plugin.app, nodes, el, ctx);
 }
 
 function unescapePipe(s: string): string {
